xtc: Refactor, own module for each subcommand
Move 'source' and 'update' subcommands to their own module as well.
