+++ tools/xtc/src/sourcegen.h
@@ -1,0 +1,6 @@
+#ifndef XTC_SOURCEGEN_H
+#define XTC_SOURCEGEN_H
+
+int dosource(int argc, char **argv);
+
+#endif
