+++ tools/xtc/src/updater.h
@@ -1,0 +1,6 @@
+#ifndef XTC_UPDATER_H
+#define XTC_UPDATER_H
+
+int doupdate(int argc, char **argv);
+
+#endif
