+++ tools/xtc/src/xtc.mk
@@ -1,5 +1,7 @@
 xtc_MODULES=	compiler \
 		deffile \
+		sourcegen \
+		updater \
 		util \
 		xtc
